requirements.txt
streamlit>=1.28.0
numpy>=1.26.0
matplotlib>=3.8.0
plotly>=5.17.0
sympy>=1.12
pandas>=2.1.0
pillow>=10.0.0
scipy>=1.11.0
networkx>=3.1
reportlab>=4.0.0
bcrypt>=4.0.0
pyjwt>=2.8.0
python-dotenv>=1.0.0